+++ asyncapi/exceptions.py
@@ -1,6 +1,18 @@
-class DoodleError(Exception):
+class AsyncApiError(Exception):
     ...
 
 
-class ReferenceNotFoundError(DoodleError):
+class ReferenceNotFoundError(AsyncApiError):
+    ...
+
+
+class InvalidChannelError(AsyncApiError):
+    ...
+
+
+class OperationIdNotFoundError(AsyncApiError):
+    ...
+
+
+class ChannelOperationNotFoundError(AsyncApiError):
     ...