+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Create publish and listen methods for api
+
 ###  - 2020-07-02
 
  - Create Api object from asyncapi spec